# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API Super Tienda",
    "contact": {
      "name": "Super    Tienda",
      "url": "https://www.superTienda.com",
      "email": "superTienda@gmail.com"
    }
  },
  "components": {
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "module": {
            "type": "string",
            "default": ""
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RolePermission"
            }
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "parentId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "sortOrder": {
            "type": "integer",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "default": 0
          },
          "updatedBy": {
            "type": "integer",
            "default": 0
          },
          "deletedBy": {
            "type": [
              "integer",
              "null"
            ]
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/definitions/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "childrens": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Category"
            }
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Product"
            }
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "businessId": {
            "type": "integer",
            "default": 1
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": [
              "string",
              "null"
            ]
          },
          "documentType": {
            "type": "string",
            "default": "CI",
            "enum": [
              "NIT",
              "CI",
              "PASSPORT",
              "FOREIGN_ID"
            ]
          },
          "documentNumber": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "city": {
            "type": [
              "string",
              "null"
            ]
          },
          "department": {
            "type": [
              "string",
              "null"
            ],
            "default": "LA_PAZ",
            "enum": [
              "LA_PAZ",
              "COCHABAMBA",
              "SANTA_CRUZ",
              "ORURO",
              "POTOSI",
              "CHUQUISACA",
              "TARIJA",
              "BENI",
              "PANDO"
            ]
          },
          "country": {
            "type": "string",
            "default": "Bolivia"
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "creditLimit": {
            "type": [
              "number",
              "null"
            ]
          },
          "currentBalance": {
            "type": "number",
            "default": 0
          },
          "loyaltyPoints": {
            "type": "integer",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "default": 0
          },
          "updatedBy": {
            "type": "integer",
            "default": 0
          },
          "deletedBy": {
            "type": [
              "integer",
              "null"
            ]
          },
          "business": {
            "$ref": "#/definitions/Business"
          },
          "carts": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Cart"
            }
          },
          "sales": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Sale"
            }
          },
          "customerDebts": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/CustomerDebt"
            }
          }
        },
        "required": [
          "firstName"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/categories": {
      "get": {
        "summary": "Listar todas las categorías",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nueva categoría",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada"
          },
          "400": {
            "description": "Categoría duplicada"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "summary": "Obtener categoría por ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar categoría",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar categoría",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Categoría eliminada exitosamente"
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "summary": "Listar todos los clientes",
        "tags": [
          "Customers"
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes"
          }
        }
      },
      "post": {
        "summary": "Crear nuevo cliente",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente creado"
          },
          "400": {
            "description": "NIT o Email duplicado"
          }
        }
      }
    },
    "/api/customers/search": {
      "get": {
        "summary": "Filtrar clientes por documento",
        "description": "Filtra clientes por tipo y/o número de documento (búsqueda parcial)",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "documentType",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "CI",
                "NIT",
                "PASSPORT",
                "FOREIGN_ID"
              ]
            },
            "description": "Tipo de documento (opcional)"
          },
          {
            "in": "query",
            "name": "documentNumber",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Número de documento (búsqueda parcial, opcional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes que coinciden con los filtros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Al menos un filtro debe ser proporcionado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Obtener cliente por ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar cliente",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar cliente",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente eliminado exitosamente"
          }
        }
      }
    },
...
    "/api/permissions": {
      "get": {
        "summary": "Obtener todos los permisos",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de permisos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo permiso",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permiso creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación o datos inválidos"
          }
        }
      }
    },
    "/api/permissions/{id}": {
      "get": {
        "summary": "Obtener un permiso por ID",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permiso encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un permiso existente",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permiso actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un permiso por ID",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permiso eliminado correctamente"
          },
          "404": {
            "description": "Permiso no encontrado"
          }
        }
      }
    },
...
  },
...
}